refactor relay init / network setup
