+++ lib/relay.ts
@@ -1,43 +1,31 @@
 import { useMemo } from 'react';
-import { Environment, Network, Observable, RecordSource, Store } from 'relay-runtime';
+import { Environment, RecordSource, Store } from 'relay-runtime';
 import { RecordMap } from 'relay-runtime/lib/store/RelayStoreTypes';
 
-import fetchMultipart from 'fetch-multipart-graphql';
+import { network } from './relayNetwork';
+
+const STORE_ENTRIES = 150;
+const STORE_CACHE_RELEASE_TIME = 2 * 60 * 1000; // 2 mins
 
 let relayEnvironment: any;
 
-// Define a function that fetches the results of an operation (query/mutation/etc)
-// and returns its results as a Promise
-function fetchQuery(operation, variables) {
-    return Observable.create((sink) => {
-        fetchMultipart(process.env.NEXT_PUBLIC_RELAY_ENDPOINT ?? '/api/graphql', {
-            method: 'POST',
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-                query: operation.text,
-                variables,
-            }),
-            credentials: 'same-origin',
-            onNext: (parts) => sink.next(parts),
-            onError: (err) => sink.error(err as Error), //wrap it later
-            onComplete: () => sink.complete(),
-        });
-    });
-}
-
 function createEnvironment() {
+    const source = new RecordSource();
+    const store = new Store(source, {
+        gcReleaseBufferSize: STORE_ENTRIES,
+        queryCacheExpirationTime: STORE_CACHE_RELEASE_TIME,
+    });
     return new Environment({
-        // Create a network layer from the fetch function
-        network: Network.create(fetchQuery),
-        store: new Store(new RecordSource()),
+        network,
+        store,
+        log(_event) {
+            // comment-out for logging:
+            // console.log(_event);
+        },
     });
 }
 
 export function initEnvironment(initialRecords?: RecordMap) {
-    // Create a network layer from the fetch function
     const environment = relayEnvironment ?? createEnvironment();
 
     // If your page has Next.js data fetching methods that use Relay, the initial records
@@ -47,6 +35,7 @@ export function initEnvironment(initialRecords?: RecordMap) {
     }
     // For SSG and SSR always create a new Relay environment
     if (typeof window === 'undefined') return environment;
+
     // Create the Relay environment once in the client
     if (!relayEnvironment) relayEnvironment = environment;
